Fixed bugs regarding transitions and new scripts, and Nomina Date filter
Fixed bug that caused errors when loading a script in the newContainer after transition (does not work on reload yet). Fixed Nomina filters that were pointing to Empleados controller instead of Nomina. Added PagedList to missing Index on both Empleado and Nomina (role based).  Fixed a bug when creating Nomina that didn't set the Fecha to the sent value. Made it so that you can delete Nominas with a confirm dialog (using ajax). Redesigned Nomina Create and Edit and added missing stuff.